Rename search::NumericAttribute::Equal to search::attribute::NumericMatcher. Rename search::NumericAttribute::Range to search::attribute::NumericRangeMatcher.
