+++ searchlib/src/vespa/searchlib/attribute/numericbase.cpp
@@ -1,7 +1,7 @@
 // Copyright Yahoo. Licensed under the terms of the Apache 2.0 license. See LICENSE in the project root.
 
 
-#include "numericbase.hpp"
+#include "numericbase.h"
 
 #include <vespa/log/log.h>
 LOG_SETUP(".searchlib.attribute.numericbase");
@@ -30,19 +30,5 @@ NumericAttribute::load_posting_lists_and_update_enum_store(enumstore::Enumerated
 {
     LOG_ABORT("Should not be reached");
 }
-
-template class NumericAttribute::Range<int8_t>;
-template class NumericAttribute::Range<int16_t>;
-template class NumericAttribute::Range<int32_t>;
-template class NumericAttribute::Range<int64_t>;
-template class NumericAttribute::Range<float>;
-template class NumericAttribute::Range<double>;
-
-template class NumericAttribute::Equal<int8_t>;
-template class NumericAttribute::Equal<int16_t>;
-template class NumericAttribute::Equal<int32_t>;
-template class NumericAttribute::Equal<int64_t>;
-template class NumericAttribute::Equal<float>;
-template class NumericAttribute::Equal<double>;
 
 }
