+++ searchlib/src/vespa/searchlib/attribute/numericbase.h
@@ -25,68 +25,6 @@ protected:
     virtual void load_posting_lists_and_update_enum_store(enumstore::EnumeratedPostingsLoader& loader);
     bool onAddDoc(DocId) override { return true; }
 
-    template<typename T>
-    class Equal
-    {
-    private:
-        T _value;
-        bool _valid;
-    protected:
-        Equal(const QueryTermSimple &queryTerm, bool avoidUndefinedInRange);
-        bool isValid() const { return _valid; }
-        bool match(T v) const { return v == _value; }
-        Int64Range getRange() const {
-            return Int64Range(static_cast<int64_t>(_value));
-        }
-    };
-
-    template<typename T>
-    class Range
-    {
-    protected:
-        T _low;
-        T _high;
-    private:
-        bool _valid;
-        int _limit;
-        size_t _max_per_group;
-    protected:
-        Range(const QueryTermSimple & queryTerm, bool avoidUndefinedInRange=false);
-        Int64Range getRange() const {
-            return Int64Range(static_cast<int64_t>(_low), static_cast<int64_t>(_high));
-        }
-        bool isValid() const { return _valid; }
-        bool match(T v) const { return (_low <= v) && (v <= _high); }
-        int getRangeLimit() const { return _limit; }
-        size_t getMaxPerGroup() const { return _max_per_group; }
-
-        template <typename BaseType>
-        search::Range<BaseType>
-        cappedRange(bool isFloat)
-        {
-            BaseType low = static_cast<BaseType>(_low);
-            BaseType high = static_cast<BaseType>(_high);
-
-            BaseType numMin = std::numeric_limits<BaseType>::min();
-            BaseType numMax = std::numeric_limits<BaseType>::max();
-
-            if (isFloat) {
-                if (_low <= (-numMax)) {
-                    low = -numMax;
-                }
-            } else {
-                if (_low <= (numMin)) {
-                    low = numMin + 1; // we must avoid the undefined value
-                }
-            }
-
-            if (_high >= (numMax)) {
-                high = numMax;
-            }
-            return search::Range<BaseType>(low, high);
-        }
-
-    };
 public:
     DECLARE_IDENTIFIABLE_ABSTRACT(NumericAttribute);
 };
